Move code to NAudioMicrophone, NAudioSpeaker
