--- MyPal.ClassLibrary/Platforms/Windows/PlatformClass1.cs
@@ -1,7 +1,0 @@
-﻿namespace MyPal.ClassLibrary
-{
-    // All the code in this file is only included on Windows.
-    public class PlatformClass1
-    {
-    }
-}
