--- MyPal.ClassLibrary/Platforms/iOS/PlatformClass1.cs
@@ -1,7 +1,0 @@
-﻿namespace MyPal.ClassLibrary
-{
-    // All the code in this file is only included on iOS.
-    public class PlatformClass1
-    {
-    }
-}
